+++ src/pages/404/NotFoundPage.jsx
@@ -1,14 +1,18 @@
 import React from 'react'
-import { Container, Img, Text, SearchLink } from './NotFoundPage.style';
+import { NotFoundContainer, Img, BackLink } from './NotFoundPage.style';
 import iconOgucatNotfound from 'assets/images/icon_ogucat_notfound.png'
+import { Text } from '../../components/home/NoneFeed.style';
+import { useNavigate } from 'react-router-dom';
 
 export default function NotFoundPage() {
+  const navigate = useNavigate();
+
   return (
-    <Container>
+    <NotFoundContainer>
       <Img src={iconOgucatNotfound} alt="404 오류. 찾을 수 없는 페이지" />
       <Text>페이지를 찾을 수 없습니다. :(</Text>
-      <SearchLink href = "history.back();">이전 페이지</SearchLink>
-    </Container>
+      <BackLink onClick={() => {navigate(-1)}}>이전 페이지</BackLink>
+    </NotFoundContainer>
   );
 }
