+++ src/router/AppRouter.jsx
@@ -1,3 +1,4 @@
+import NotFoundPage from 'pages/404/NotFoundPage';
 import HomePage from 'pages/HomePage/HomePage';
 import MainLoginPage from 'pages/LoginPage/MainLoginPage';
 import React from 'react';
@@ -7,8 +8,9 @@ export default function AppRouter() {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/home" element={<HomePage />} />
         <Route path="/" element={<MainLoginPage />} />
+        <Route path="/*" element={<NotFoundPage />} />
+        <Route path="/home" element={<HomePage />} />
       </Routes>
     </BrowserRouter>
   );
